# CASMINST-5159: Speed up Goss test result reporting