--- a/goss-testing/automated/livecd-provisioning-checks
+++ b/goss-testing/automated/livecd-provisioning-checks
@@ -24,7 +24,7 @@
 
 # GOSS_BASE should already be set on the PIT node, but just in case we
 # will set it here
-export GOSS_BASE=${GOSS_BASE-"/opt/cray/tests/install/livecd"}
+export GOSS_BASE=${GOSS_BASE:-"/opt/cray/tests/install/livecd"}
 
 source "${GOSS_BASE}/automated/run-ncn-tests.sh"