## Summary and Scope

A few of the Goss automated scripts use Bash scripting and a lot of jq commands to parse the results and display them. This is very slow and can take longer than the tests themselves, more than doubling the time of the script execution.

This PR creates a Python script which replaces that code. The script makes a GET request to a Goss endpoint, or reads Goss JSON results from stdin (or from a file). Either way, it then parses the JSON results and presents them in the same way that the current scripts do. Just much faster (see testing section for details).

[CASMINST-5159](https://jira-pro.its.hpecorp.net:8443/browse/CASMINST-5159) is phase 1 of a multi-part plan. The next stage is [CASMINST-5161](https://jira-pro.its.hpecorp.net:8443/browse/CASMINST-5161), to modify our scripts which make calls to multiple goss endpoints so that they can present a grand total result summary at the end of the output, thus removing the need for our current "grep for the total lines" stuff in our docs. The goal here would also be to only output failures (if any) and a summary of execution -- not the current page after page of test successes. That output would instead be sent to a log file.

The final phase is [CASMINST-5162](https://jira-pro.its.hpecorp.net:8443/browse/CASMINST-5162), to, if our tests don't have conflicts that prevent it, hit  the test endpoints in parallel on the nodes.

But for this PR, it just replaces the JSON parsing with something much faster.

## Testing

Three files contained this parsing code:
livecd-preflight-checks
livecd-provisioning-checks
run-ncn-tests.sh

I tested the two livecd scripts on the surtur PIT node. 

I tested the run-ncn-tests.sh changes by calling the ncn-healthcheck script on rocket, because it in turn calls the three ncn-healthcheck-master, -storage, and -worker scripts, and each of those call the modified function in run-ncn-tests.sh. 

I ran the scripts both with and without this PR in place. The results were the same for both in terms of the tests being executed and the output reported (as in, no problems were introduced).  However, the speeds were improved:

Script | Time without PR | Time with PR 
------ | ------------------ | ---------------
livecd-preflight-checks | 54.364s | 17.516s
livecd-provisioning-checks | 3.286s | 0.398s
ncn-healthcheck | 5m27.408s | 3m10.443s

In addition, I made sure that I tested cases where there were test failures, to ensure that they were properly handled.

## Risks and Mitigations

The only change is to the logic for parsing the test output. Low risk.

## Pull Request Checklist

- [X] License file intact
- [X] Target branch correct
- [X] Testing is appropriate and complete, if applicable